% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printMedPathCoeff.R
\name{printMedPathCoeff}
\alias{printMedPathCoeff}
\title{printMedPathCoeff}
\usage{
printMedPathCoeff(theta)
}
\arguments{
\item{theta}{List of mediation path coefficients for each PDM}
}
\value{
A printed, formatted table of the mediation path coefficient values
for a, b, ab, c' for each calculated PDM
}
\description{
Prints out mediation path coefficients in a formatted table
}
