% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceMedDimension.R
\name{reduceMedDimension}
\alias{reduceMedDimension}
\title{reduceMedDimension}
\usage{
reduceMedDimension(treatment, outcome, mediator, exVar = 0.9)
}
\arguments{
\item{treatment}{vector (N x 1) Treatment vector for N Subjects}

\item{outcome}{vector (N x 1) Outcome vector for N Subjects}

\item{mediator}{matrix (N x p) Mediator matrix for N subjects and p features
(Assumed p >> N)}

\item{exVar}{Percentage of Mediator matrix variance user wants SVD to explain.
Default is 90\%}
}
\value{
A list
\itemize{
\item X - vector (N x 1) Treatment vector for N subjects
\item Y - vector (N x 1) Outcome vector for N subjects
\item M - matrix (N x p) Original Mediator matrix for N subjects and
p features
\item M_reduced - matrix (N x b) Mediator matrix with reduced dimension
space of b features where N > b
\item tLoadingMatrix - matrix (p x b)' Eigenvector space from SVD that
maps M_reduced back to the original Mediator matrix
\item exVar - Percentage of Mediator matrix variance user wants SVD to
explain
}
}
\description{
Reduces dimensions of mediator space using Singular
Value Decomposition (SVD). Moves treatment, outcome, and mediator
data into a list to be used as an input in later bioPDM steps.
}
