% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveFilesForVisPDM.R
\name{saveFilesForVisPDM}
\alias{saveFilesForVisPDM}
\title{saveFilesForVisPDM}
\usage{
saveFilesForVisPDM(data_list, pdm_result_list, saveDir = NULL)
}
\arguments{
\item{data_list}{List containing X, Y, M, M_reduced, tLoadingMatrix,
and exVar elements. Output from reduceMedDimension()}

\item{pdm_result_list}{List output from getDirectionsOfMed()}

\item{saveDir}{Directory for which the user wants to save results to}
}
\value{
.csv files. *PDMs_X_Y.csv files contain the PDM, X, and Y values for
each subject. *PDM_feature_loading.csv files contain the feature weights for
each PDM
}
\description{
Saves .csv files that can be uploaded into visPDM app.
}
